+++ paper-io/MainWindow.xaml.cs
@@ -8,7 +8,6 @@ using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
@@ -23,24 +22,14 @@ namespace paper_io
         public MainWindow()
         {
             InitializeComponent();
-            Player player = new Player(0, 0, Color.FromArgb(220, 220, 220, 220));
-
-
         }
+
         private void StartButton(object sender, RoutedEventArgs e)
         {
-            Game game = new Game(Convert.ToByte(EnterField.Text));
 
-
-            //for (int i = 0; i < game.players.GetLength(0); i++)
-            //{
-            //    for (int j = 0; j < game.players.GetLength(1); j++)
-            //    {
-            //        MessageBox.Show(Convert.ToString(game.players[i, j]));
-            //    }
-            //}
         }
-        private void EnterField_TextChanged(object sender, TextChangedEventArgs e)
+
+        public void EnterField_TextChanged(object sender, TextChangedEventArgs e)
         {
             string result = "";
             foreach (char item in EnterField.Text)
@@ -49,45 +38,6 @@ namespace paper_io
                     result += item;
             }
             EnterField.Text = result;
-        }
-    }
-
-    class Game
-    {
-        /// <summary>
-        /// Количество игроков
-        /// </summary>
-        public Player[,] Players { get; set; }
-        public Game(byte players)
-        {
-            Players = new Player[players * 10, players * 10];
-        }
-    }
-
-    class Player
-    {
-        /// <summary>
-        /// Координата расположения игрока по X
-        /// </summary>
-        public int X { get; set; }
-        /// <summary>
-        /// Координата расположения игрока по Y
-        /// </summary>
-        public int Y { get; set; }
-        /// <summary>
-        /// Цвет игрока
-        /// </summary>
-        public Color ColorOfPlayer { get; set; }
-        /// <summary>
-        /// Жизнь игрока
-        /// </summary>
-        private readonly bool life = true; 
-
-        public Player(int x, int y, Color color)
-        {
-            X = x;
-            Y = y;
-            ColorOfPlayer = color;
         }
     }
 }
